jquant/requirements.txt
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
idna==2.9
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
numpy==1.18.1
pandas==1.0.3
pycparser==2.20
pyOpenSSL==19.1.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
six==1.14.0
TA-Lib==0.4.18
urllib3==1.25.8
websocket-client==0.57.0
win-inet-pton==1.1.0
wincertstore==0.2
